requirements.txt
aiobotocore==2.13.1
aiohttp==3.9.5
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.13.2
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
apache-airflow==2.9.2
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.14.1
apache-airflow-providers-fab==1.2.0
apache-airflow-providers-ftp==3.10.0
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.1
apispec==6.6.1
appdirs==1.4.4
argcomplete==3.4.0
asgiref==3.8.1
async-timeout==4.0.3
asyncssh==2.15.0
atpublic==4.1.0
attrs==23.2.0
Babel==2.15.0
billiard==4.2.0
blinker==1.8.2
boto3==1.34.131
botocore==1.34.131
cachelib==0.9.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
cloudpickle==3.0.0
cmdstanpy==1.2.4
colorama==0.4.6
colorlog==4.8.0
configobj==5.0.8
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.2.1
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==42.0.8
cycler==0.12.1
Deprecated==1.2.14
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
docutils==0.21.2
dpath==2.2.0
dulwich==0.22.1
dvc==3.51.2
dvc-data==3.15.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-s3==3.2.0
dvc-studio-client==0.21.0
dvc-task==0.4.0
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
google-re2==1.1.20240702
googleapis-common-protos==1.63.2
grandalf==0.8
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.64.1
gto==1.7.1
gunicorn==22.0.0
h11==0.14.0
holidays==0.52
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hydra-core==1.3.2
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflection==0.5.1
iterative-telemetry==0.0.8
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kombu==5.3.7
lazy-object-proxy==1.10.0
lightgbm==4.4.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.9.1
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
mlflow==2.14.2
more-itertools==10.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
ordered-set==4.1.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pendulum==3.0.0
pillow==10.4.0
platformdirs==3.11.0
pluggy==1.5.0
prison==0.2.1
prompt_toolkit==3.0.47
prophet==1.1.5
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.9
pyarrow==15.0.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydot==2.0.0
pygit2==1.15.1
Pygments==2.18.0
pygtrie==2.5.0
PyJWT==2.8.0
pyparsing==3.1.2
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
querystring-parser==1.2.4
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.2
rpds-py==0.19.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3fs==2024.6.1
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.13.1
scmrepo==3.3.6
semver==3.0.2
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
sqltrie==0.11.0
stanio==0.5.0
starlette==0.37.2
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.14.2
tomlkit==0.12.5
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==1.26.19
uvicorn==0.30.1
uvloop==0.19.0
vine==5.1.0
voluptuous==0.15.2
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.19.2

